+++ src/parser/mod.rs
@@ -2,3 +2,6 @@ pub mod recursive_descent_parser;
 mod recursive_descent_parser_spec;
 mod recursive_descent_parser_transitions;
 mod recursive_descent_parser_transitions_spec;
+mod failed_production;
+mod failed_state;
+mod parse_result;