+++ src/CHANGELOG.md
@@ -19,7 +19,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## [UNRELEASED]
+
+### Added
+- Added `FailedProduction` struct.
+- Added `FailedState` struct.
+- Added `ParseResult` type.
+
+### Changed
+- [BC] Updated `RecursiveDescentParser.parse_from_tokens` to return a `Result` instead on an option.
+
+### Fixed
+- Fixed `RecursiveDescentParser.parse_from_tokens` bug which caused the parser to parse invalid tokens when multiple productions hit at least a production symbol.
+
+
+
+
 ## v0.2.0
+
+### Changed
 - [BC] Updated `recursive_descent_parser` to receive an iterator of `Token<TLex, TSyntax>>`.